Rewrite of Simulator to use Theano-like operators.
The goal of this rewrite was to simplify the relationship between signals and
buffers in the simulator, and to simplify the mental model underlying the
simulator. There are no longer different "phases" of the simulator (i.e.
Filters, Transforms, Encoders, Decoders) with their somewhat overlapping
memory access rules, instead there is just a graph of linear operators which
is sorted topologically.