Updated all examples to new syntax
They all run without exceptions on 2.6, 2.7 and 3.3.

Also now using matplotlib master branch.
Because of how tox invokes pip, using the version in PyPI
won't work, as it tries to pip install numpy and matplotlib
in the some invocation. Recent versions of matplotlib have
fixed this issue, so for the time being, we'll recommend
the latest matplotlib in requirements-test.txt so that
tox works.