Added rectified_linear example
This is an advanced example of how you can define a new neuron type
within a user script, if you want to do that kind of thing.

Mostly, it's a justification for the pluggable builder changes in
the previous commit. I don't believe that you could do this example
before the previous commit without doing some pretty hacky stuff.