Enabled builder validation and optimization
validators and optimizers get registered just like builders.
Most validators / optimizers will be on networks, but it's possible
to do either on any type of object; this is useful for doing
general validations or optimizations on all objects, or on models,
by registering the validator / optimizer to `object`, which
all classes subclass from.

Other changes:
- Made Builder.build return the model; this is useful for
  quickly building an object and getting the model back
- Added test to ensure registration works properly