Rewrote configuration system
This previous system was good, but was difficult to change on the fly,
which is necessary for setting object defaults (see PR #329).
This system should support setting object defaults cleanly,
without needing separate classes for each type that will be configured.

The most relevant changes are in `nengo/config.py`. Now, instead of
subclassing, each backend should instantiate a `Config` object
and add extra parameters to it by calling `set_param`.
`ClassParams` and `InstanceParams` are classes similar to
the old `ConfigItem`, but they are also designed to be extended live,
instead of through subclassing.
It is also created automatically by the `Config` object;
`ClassParams` and `InstanceParams` shouldn't be instantiated manually.
Instead, use the `configures` function.

Looking at the tests in `nengo/tests/test_config.py` should give
a good picture of how to use the new Config system.

Other changes:
- builder.Model keeps track of which network is the top level
  network, for possible future use
- network.config is passed in to all build methods called
  by build_network, for possible future use
- Networks get a config object upon instantiation; the config
  object is constructed in the `Network.default_config` function.
- Added a `params` property to NengoObject that returns a list
  of Parameters
- Made all appropriate NengoObject attributes `config.Parameter`s
  so that they share the implementation used by the config system
- Got rid of **kwargs in Connection.__init__
- Configs mess with pickling because of the WeakKeyDict.
  I think there are lots of reasons to not use Pickle though,
  so I got rid of saving / loading networks for now, and we should
  instead do something more intelligent in the future
- Fixed Uniform.__eq__ and Gaussian.__eq__ when comparing to
  other types of objects