Added Parameter subclasses to simplify objects
Where possible, the validation logic in objects.py has been
moved to Parameter subclasses in nengo.params, in order to
put all front-end validation in one place, play nice with the config
system, increase the amount of parameter validation that we do,
and simplify the user-facing objects.

So far, some basic types (`IntParam`, `NumberParam`, etc) are
defined, and one complicated type (`NodeOutput`) as a proof
of concept that these descriptors are sufficiently powerful.
Next step is to implement more of these.

One more thing we could do with this is to restrict the ability
to set any attribute that isn't a Parameter subclass. We discussed
doing this before, and people were into it.