Added DistributionParam
This describes a parameter that can be a nengo.utils.distributions
Distribution, or samples from some arbitrary distribution.
If it is samples, you can give it attributes that describe the
correct sample shape.

This required a few changes to the distributions; the dimensions
are passed to the sample function now instead of being stored
on the object. This makes it possible to instantiate distributions
before the build phase.