Made bias and gain into DistributionParam
Now you can set them for ensembles. Note that, as before, if you
set the gain and bias it will be used, but if you only set one
then it will ignore it and use the max_rates and intercepts.
It should probably work with only one. Also, if you give gain
and bias, you should set the max_rates and intercepts;
basically, you should have to set two or three of the four,
and Nengo should fill the rest in intelligently.
That's a todo for later though.