Added SynapseParam
For synapses! If that wasn't already obvious. A bit of an API
change here: now, setting the synapse to a float make a Lowpass
object at the frontend, rather than filling in that detail in
the builder. @hunse and I couldn't come up with a reason
for waiting to fill it in.