Force setuptools
If we can be assured that we're using setuptools,
we can make setup.py a lot more effective. This should be possible,
as the ez_setup.py script (now included) should be able to download
and install setuptools on all OSes.

A few nice things if we can rely on setuptools
- simpler setup.py
- use find_packages function instead of having to edit setup.py
  when we add a new package