Complete restructure of docs directory
In preparation for the 2.0.0 release, this is an overhaul of our docs.
I've rewritten a lot of the text and rearranged the files.

Another big change is adding in RST files for all of the examples,
which causes Sphinx to evaluate all of the examples when you build the
documentation (with `sphinx-build` or `python setup.py build_sphinx`).
The IPython notebook examples are then inserted inline in the HTML file
generated by Sphinx, along with links to the fully evaluated ipynb,
and an equivalent raw Python script.

This change means that we cannot use the ReadTheDocs service.
ReadTheDocs is not set up to run these expensive examples.
That's okay though; they are easy to run locally, and we can
use PyPI's document hosting or a local server to host the
documentation. We can either build it once every release,
or have a Travis-CI job that builds it and uploads it someplace.

However, since the ReadTheDocs theme is pretty sweet, we still
use it. It's not quite as configurable as would be ideal, so at
least for the time being, I've forked their theme as a new
nengo_sphinx_theme project, which Nengo uses. The important change
between nengo_sphinx_theme and the ReadTheDocs theme is better
monospaced fonts on Mac OS X, and the syntax highlighting in
all of the documentation matches that of IPython's, which means
that all of the documentation has a consistent style, whether it's
code examples in docstrings or IPython notebook examples.