Fail faster when making Ops
Before, make_step wouldn't be called until Simulator.__init__.
If there was an issue (e.g., shape mismatch) you would get the error
there, which is not very helpful in debugging.
Now, when you add an op to the model, it calls make_step right away
to ensure that it works. It wastes a bit of time during build, but
doesn't make any impact on simulation time, so it's worth it for
the easier debugging.