First pass at displaying spikes.
Spikes can now be displayed in javaviz if the node's spikes are probed
in the script. Currently the spiketimes are noisier than they should be,
but I'm not sure why. Job for tomorrow.

Also, currently assumes each ensemble's spikes are probed at most once,
should be made to work if the spikes are probed multiple times.