Implemented SignalView.value
Before, it was not possible to get the value of a view, only the
base. This change makes it possible to get the value of a SignalView,
which is just a view of signal.base.value. This means that it's now
also possible to set / reset signal views without changing the
entire base signal; see the changes to test_builder.py.