Fixed javaviz bug related to filenames.
For scripts where the top-level network does not have a label, the label
that nengo_gui passes to javaviz so that javaviz can load/save layouts
is based on the script's filename. But for files inside of a folder (ie
layout tests/spa.py), the filename contains a '/' which messes with
javaviz's loading/saving of layouts.