Put has_built check back in
This check does make sense, as long as the commented out bit
underneath is removed. Basically, if a build method is designed
to only happen once, it should add results to model.params.
If it shouldn't (e.g., synapses, because they build for each
connection, so it's really the (synapse, connection) tuple that
gets built), then we don't add it to model.params,
and that check passes.