Made NdarrayParam
Which DistributionParam is a subclass of now. The initial reason
for this was because Connection.eval_points doesn't actually
allow for a distribution to be passed in, though I think this is
probably not correct behavior. Regardless, it is nice to have
this separate, as Transform should probably be a subclass of this.