Exposed FunctionInfo
This is designed to make it more clear what is happening
with the FunctionParam. Instead of keeping the FunctionInfo
inside and overriding __get__ etc to return information, it
just stores and returns the FunctionInfo. The `Connection`
class is therefore modified to reflect this, as well as
some tests and builder code. In Connection, I use properties
so that existing code doesn't change.

I'm not 100% sure that this makes things more clear...
feedback welcome!