Removed NengoObjectInfo
Similar to the FunctionInfo change, except instead of exposing
NengoObjectInfo, I removed it and replaced it with some Connection
properties. This simplified params.py a bit, but perhaps made
Connection more complex.

This is an API change, because now Connection.pre and post
can be ObjViews; this didn't result in too many code changes, and is
hopefully less surprising to users, so this should be a net gain.
If a user wants a guaranteed object, they can use Connection.pre_obj
and post_obj.

I also made the ObjView a little bit more robust, by pre-caching
its size_in and size_out (and getting rid of the role stuff).
It also now has a very nice label (which is useful since
people are likely to see these labels when they print out
conn.pre and conn.post).