Added validation and more defaults to Parameter
Parameter and Default now live in params.py.

The ability to do validation is now built into params.Parameter.
Parameters can be set as optional (None is not a valid value)
or readonly (can only be set once per instance).

Additionally, multiple "default" values can be set; this is
useful for the Nengo config stack, as each config object can set
a default for each Parameter.

Tests for Parameter are moved to test_params as well, as many more
subclasses are forthcoming.