Added NdarrayParam and DistributionParam
These Parameters accept a Numpy ndarray and a
utils.distributions.Distribution, respectively.
The DistributionParam also accepts a sample with
an optionally constrained shape.

This required some significant changes to how we do distributions.
Importantly, the dimensions are passed to the sample function now
instead of being stored on the object. This makes it possible to
instantiate distributions before the build phase.

Many parameters end up being DistributionParams, including
eval_points. However, since it was also possible to set eval_points
to an integer which would be the number of eval points, we made
n_eval_points a separate IntParam.

An Ensemble's gain and bias can now be reliably set, either as an
ndarray or a distribution. These can be set after instantiation,
or passed as arguments to nengo.Ensemble().

Other miscellaneous changes:
- Removed ShapeMismatch, which was only used in SignalView.
- Removed pick_eval_points which was very short with these changes
- Product and CircularConvolution get nengo.Default instead of None
  for the encoder, as None is no longer a valid value
- spa.assoc_mem now sets properly shaped sets of eval_points

Co-authored-by: Eric Hunsberger <erichuns@gmail.com>