Added NodeOutputParam
This parameter is just for Node.output, which has some special
processing because it can be None, an ndarray, or a function.

In order to pass the Node.output unit test, NengoObjects have to
be added to networks after the __init__, in case the __init__ errors
out. If we don't do it in this way, you can get networks in
invalid states.