Hierarchical build functions
Now, the register function takes in a list of classes.
The build function will receive (and will only receive)
the model, and then instances of those types.

This makes a lot of things more straightforward, and nestable;
it's possible, for example, to have the same synapse type
share a build function when building Connections and Probes,
and it's also possible to have different build functions for
those two cases.

This means that we should endeavour to have more build functions
call other build functions, since those build functions can
be overridden at run time by the user, if they so choose.

Other changes:
- Build functions have to set themselves in model.params again.
  This was taken out because it wasn't clear what some build functions,
  like those for synapses, should set. Now we use a tuple with the
  passed arguments.
- The config is no longer passed to build functions, but is instead
  stored on the model during the network build function.
- model now has a build function, which passes itself to
  Builder.build. Saves a bunch of typing, and maybe is more clear.
- Additional validation check in LearningRuleParam to make sure
  the same rule isn't duplicated in the list, as that won't actually
  work. It does work to pass two instances of the same type,
  it just can't be the exact same object.
- test_unsupervised_learning_rule wasn't actually using its
  parametrized learning_rule arg. Now it does!