Object specific validation in param subclasses
In response to some discussion with @hunse, this commit
moves all validation specific to a NengoObject (just Connection
and Ensemble so far) to a subclass of the more general parameter.
The general parameters live in their own files (e.g.,
SolverParam in solvers.py) while the specific subclass lives in
the object's file (which has the class's name prepended; e.g.,
ConnectionSolverParam in connections.py).