Used Choice distribution wherever possible
Notably, in the Product network (and by association,
the circular convolution network), and multiplication examples.

This also fixes an issue with the Choice distribution:
the sample method should take in a `d` argument, which is the
dimensionality of the output. In most distributions, this
affects the sampling process; here, it doesn't, but it needs
to be included for compatibility. To ensure compatibility,
we instead use `d` to ensure that the options passed in are of
the correct dimensionality (tested in test_distributions.py).

Also, a quick fix to the Product network's test_sine_wave function;
the desired result was outside of the representational range
(could be as high as radius * radius), so reduced the size of the
representations to np.sqrt(radius).