Fix benchmark tests.
Tests that are run with `--benchmark` have also been failing for
some time; this fixes those tests.

In order to fix the tests, a change was made to the `int_types` in
`nengo.utils.compat`. This is due to a Python 2 / 3 difference in how
NumPy declares its integer types. On Python 3,
`isinstance(np.int64(1), int)` fails, so we add `np.integer` to the
`int_types` lists, and therefore `is_integer` works for NumPy ints
on both Python 2 and 3.