Remove FeedbackGatedMemory
In the FeedbackGatedMemory, an ensemble is put in the feedback loop
of the integrator. When input is presented, the ensemble in the
feedback loop is gated. While this can work, it requires a lot of
tweaking of `synapse` parameters and scaling constants.

It was originally implemented due to bugs in previous version of
InputGatedMemory; however, those bugs have been fixed, so users of
FeedbackGatedMemory should be able to safely switch to using
InputGatedMemory.

Other changes:
- A bit of cleanup in InputGatedMemory. Removed unnecessary
  input and output nodes.
- Refactored test_workingmemory.py and shortened the simulation time