Allow more flexible signal reshaping
Apparently the only reason we limited reshaping was to avoid
non-contiguous signals. This commit ensures that all signals
are C-contiguous numpy arrays, and therefore we allow any reshaping.

Note that there's a bit of weird edge case for scalars and the use
of ascontiguousarray; I filed a bug against Numpy for this, and
apparently there's no particularly good reason for it;
see https://github.com/numpy/numpy/issues/5300.

Also added the ability to use -1 when reshaping, like numpy arrays.
This ability is tested in test_builder.