Add IPython2ProgressBar
This is used by default if we can detect that we're in the IPython
notebook, and we have widgets available.

Added some utility functions to utils.ipython to support this;
specifically, `in_ipynb`, which checks if we're in the notebook
environment, and `has_ipynb_widgets`, which checks if the IPython
notebook environment supports Widgets (i.e., is IPython>=2.0).

Additionally, the `utils.ipython` module can now be imported even
if IPython isn't installed. Since it's imported by `utils.process`,
which is imported by `simulator`, we have to do this to avoid
a hard dependency on IPython.