Fix setup.py not installing NumPy
It seems to install just fine if NumPy is in `setup_requires`
rather than `install_requires`. I'm not sure if NumPy should be
removed from `install_requires`, but I feel like best not to touch
it in case something reads `install_requires` but not `setup_requires`.

Tested on Mac OS X 10.10 and setuptools 2.2. Doesn't install NumPy
twice, but leaves an annoying `.egg` folder in the Nengo directory
(which can be removed with no ill-effects).