Add Izhikevich neuron type
Unlike the Sigmoid and RectifiedLinear types, the Izhikevich neuron
is not really suitable to be added to the `nl` and `nl_nodirect`
fixtures; while it is possible to decode functions from Izhikevich
ensembles, it requires a lot more tweaking and care. Instead, there
is an explanatory notebook, as well as a short test that ensures that
we can simulate (without crashing) several sets of commonly used
Izhikevich model parameters. Note that without the tweaks that are
commented on in `Izhikevich.step_math`, even this simple test might
blow up.