Store cache files as NCO.
Store the pickle part and array part of the cache data in a single
"Nengo cache object" file. That reduces overhead due to a lot of small
pickle files. Also, it ensures some more consistency in the cache
because it is not possible anymore that only the pickle or the array
file gets lost. That again frees us from checking for orphans which
gives a slightly better performance.