Removed `utils.functions.whitenoise`
And used the `processes.WhiteNoise` instead. To make this relatively
easy from the end-user perspective, I added an `f` method to
`StochasticProcess` that returns a function that can be used as a
Node output. This basically just wraps the sample function and ignores
the incoming `t`.

Note that when used as a node function, processes maintain state even
if the simulation is reset. This is skirted around in `test_reset`
by using a white noise signal with half the duration of the simulation
so that it repeats. In general, however, we can't expect that node
functions will reset when the simulation resets.