Parameters default to being Unconfigurable
Specifically, the general Parameters defined in `params.py`
now default to being Unconfigurable. This makes them easy to
use in classes that are not managed by the config system.
Parameters that are configurable need a default,
so this does not affect existing code.

I also allowed readonly parameters to have a default, since
there is no reason why this can't happen (though it typically
doesn't).