Modified SPA similarity functions slightly
The version in `spa` now optionally takes in a vocab, and defaults
to the probe target's vocab if none is passed in. This mean that it
is not necessary to ever use nengo.spa.similarity if you have a
`SPA` instance.

`nengo.spa.similarity`, on the other hand, was made more generic;
now, it doesn't take in the probe, just an ndarray of data as the
first argument. Previous it used the probe to get the vocab, but
that isn't possible anymore, so this version should be easier to use.
Modified `test_assoc_mem` accordingly, which was the only direct use
in Nengo to this point.

Minor change:
- imported `enable_spa_params` into `nengo.spa` namespace