Better printing for config objects
Added str and repr for all classes in `nengo.config`: `Config`,
`ClassParams`, and `InstanceParams`. The implementations are as
discussed in #529, except that we only print the configured
parameters (no defaults).

To view defaults, I added `Config.all_defaults`, which details
the defaults for all parameters in the current context.
This should be the only way to query the current defaults;
the `str` and `repr` for `ClassParams` now only lists the
actually configured elements, if any exist.

This also removed the test for the config string, as that's a pain
to maintain, and is being displayed in the upcoming notebook, so we
will still be notified if it breaks. Also an additional test was added
to ensure that deleting configured items uses the correct fallback.