Added heuristic for rasterplot marker size
The default markersize is based on the size of the axes divided
by the number of neurons. This heuristic works well as long as
the user does not resize the plot (there's really nothing we can
do in that case except use `plt.eventplot`, which has performance
issues). We default to using `plt.plot` and the heuristic,
but users can still specify to use `plt.eventplot` if they want.

Using plot closely matches eventplot for all values of `n_neurons`,
so this is used by default now for performance reasons
(PDFs generated by `eventplot` can get very large).

Other changes:
- Implemented a simple `Plotter.Mock.__mul__` method to avoid an
  error in `test_izhikevich` when `NENGO_TEST_PLOT != 1`