ALIF ss rate estimation/gain_bias + rate unit test
Previously, the AdaptiveLIF neuron relied on LIFRate's rate and
gain_bias functions. This commit gives the AdaptiveLIF its own
rate and gain_bais functions that more accurately estimate the
neuron's steady state rate and use that to set the gain and biases.

Getting the steady state rates and gains and biases correct is
necessary for accurate decoding.