decay adaptive feedback increment before LIF step
The adaptive feedback is modeled by a first order synapse.
It is subtracted from the input current passed to the LIF model.
The order in which we increment and decay the adaptive feedback slightly
changes the neuron dynamics. You can either

Increment the state with the spikes and then decay the state
or
Decay the state and then increment with the spikes

If you increment after decaying, then the neuron will have a stronger feedback.
If you increment before decaying, then the neuring will have a weaker feedback.

Which scheme to use?
This commit changes it to incrementing before decaying. Why?
Consider a decaying exponential over a fixed time period.
If you received this decaying exponential as input, would the values you see
be closer to the initial value of the decaying exponential or the value of
the decaying exponential at the end of the fixed time period?
You would see more values closer to the value of the decaying exponential at
the end of the fixed time period. Therefore, this commit switches the order
so that the increment is decayed before being used by the LIF neuron.