add LIF step_math_noclip for adaptive neuron
The adaptive LIF neuron model does not clip voltages below 0 to 0.
This is to make the steady state estimation of the adaptive LIF neuron model
consistent with the implemented dynamical model.

When the feedback strength is stronger than the input current, the adaptive
neuron model voltage will go negative. Accounting for clipping would proably
make it intractable to compute the the ss firing rate (this is an opinion and
could be shown to be incorrect).