Error connects to LearningRule
Connect the error directly into a learning rule (as discussed in #632),
instead of making a modulatory connection. This is cleaner and clearer.

Learning rules are now built by their parent connections, and since
a connection to a learning rule must added to a network after
the connection containing the learning rule, the learning rule
will always be built when the builder for the error connection into
said learning rule is called.

This fixes #632 (slicing `post` in learned connections). The test
`test_learning_rules.py:test_pes_decoders_multidimensional` has been
updated to test this feature.