Better, faster .travis.yml
The goal here is for TravisCI to build and test Nengo faster.
The two most significant ways in which we do that are to
1) use `sudo: false` to enable container-based builds as described in
http://docs.travis-ci.com/user/workers/container-based-infrastructure/
and 2) use Miniconda to quickly download and install binaries in
the home directory, so that we don't need to use `sudo`, and
don't have to build anything from source.

Other minor changes:
- Use language=c for slightly faster builds, as TravisCI does some
  virtualenv stuff when language=python, and we don't need it.
  Putting the Python information in `env` also makes it more clear
  what's being tested.
- Ensure coverage picks up the right files
- Add scikit-learn to optional dependencies
- Set `zip_safe=False` in setup.py to suppress many useless warnings