Add analytics fixture
This allows to store additional data in test. A use case would be in
benchmarking and comparing different implementations/revisions. To
store analytics data add the following to your tests:

    def test_something(analytics):
        # code
        analytics.add_data('some_key', np.array([1, 2]), 'description')

You also have to set the NENGO_ANALYTICS environment variable to 1. The
data will be stored in the nengo.simulator.analytics subdirectory. Two
files will be generated for each test. The txt file contains the
description of the stored data and the npz file the actual data. It can
be loaded with numpy.load.