WIP: Made probes an operator
This makes probes a little less "special", in that they are dealt
with mainly by the builder instead of the simulator.

I also added a ProbeOutput class; this is the beginnings of specifying
probe outputs as described in #613. Currently, we only have ProbeBuffer,
which is what we had before, where probe output is stored in memory.

TODO: still an error in learning where the operator reads before
update has been called on the transform.