Ignore PEP8 error 731
This error states that lambda expressions should not be assigned;
instead, they should be done with a `def`. We could change these,
but we only do it when doing so would save a few lines, and the
function is sufficiently simple.

Also fix error 402: module level import not at top of file
in `test_examples.py`.

Note that `flake8` by default ignores E123,133,226,241,242
so we also ignore these, for now.