Clean up PreserveValue operator
This operator is a bit hacky, but after some experimenting, this is
essentially the cleanest way to ensure that all learning operations
are applied correctly.

Increments and reads happen *after* a signal is set. By setting the
transform with PreserveValue, we allow increments to be done by
learning rules. The only other alternative is to update it, which
happens *last*. It's not obvious why one learning rule would be
applied before another, so by using PreserveValue, we do the delta
computations on the previous timestep, and then increment the transform
on the current timestep's 'increments' phase. The transform isn't
updated, but that's okay because it's set.