Fix test running in IPython 3.0
IPython 3.0 switches to the new notebook format 4. Some of the code
we were using relies on some format 3 things; this fixes those issues.
Our tests should run on notebook formats 3+; we'll need to make sure
we don't commit any notebooks with format 4 as that requires us
to depend on IPython>=2.4.1.