Define py.test options in separate module.
With defining the options in their own module they can be loaded early
by py.test with the `-p` option. Passing the `conftest.py` file to the
`-p` option leads to problems as the file would be loaded a second time
later on.

When running the tests from within the nengo directory tree everything
should work as before, but additionally a simple `py.test` should work
(without any further arguments). This also makes is possible to run
the tests of an installed Nengo package from anywhere in the file
system with `py.test --pyargs nengo -p nengo.tests.options`.

Fixes #704.